Change test to demonstrate type conversion bug
https://github.com/AssemblyScript/assemblyscript/issues/429
