+++ tests/near-bindgen/test.sh
@@ -2,7 +2,8 @@
 set -e
 set -x
 BASEDIR=../..
-$BASEDIR/bin/asc main.ts --outFile main.wat --nearFile main_near.ts
+$BASEDIR/bin/asc model.ts --nearFile model_near.ts
+$BASEDIR/bin/asc main.ts --nearFile main_near.ts
 $BASEDIR/bin/asc main_near.ts -o main_near.wat
 $BASEDIR/bin/asc test.ts -o test.wasm --textFile test.wat
 prettier --parser typescript --write main_near.ts
